cambios de las carpetas data y components graphs
