--- eduardo/df_benf_depto.csv
@@ -1,25 +1,0 @@
-,DEPARTAMENTO,cant_benef
-0,ANTIOQUIA,20641
-1,ATLANTICO,90253
-2,BOGOTA D.C.,123019
-3,BOLIVAR,25099
-4,BOYACA,11399
-5,CALDAS,7836
-6,CAQUETA,20998
-7,CASANARE,2629
-8,CAUCA,13034
-9,CESAR,26070
-10,CUNDINAMARCA,408
-11,HUILA,42826
-12,MAGDALENA,26814
-13,META,10641
-14,NARIÑO,33329
-15,NO DISPONIBLE,116825
-16,NORTE DE SANTANDER,26031
-17,PUTUMAYO,46784
-18,QUINDIO,7153
-19,RISARALDA,30946
-20,SANTANDER,17307
-21,SUCRE,21302
-22,TOLIMA,41313
-23,VALLE DEL CAUCA,39774
