--- eduardo/df_prediccion_benef.csv
@@ -1,2 +1,0 @@
-,Valor_Mensual,edad_aprox,Sexo_F,Sexo_M,Nombre_UEN_POLICIA NACIONAL,Nombre_UEN_ELECTRIFICADORAS,Nombre_UEN_EMPRESARIAL,Nombre_UEN_FUERZAS MILITARES,Sucursal_Outsourcing,Sucursal_Cali,Sucursal_Melgar,Sucursal_Florencia,Sucursal_Ibague,Sucursal_Villavicencio,Sucursal_Sincelejo,Sucursal_Pasto,Sucursal_Bogota,Sucursal_Manizales,Sucursal_Tunja,Sucursal_Medellin,Sucursal_Popayan,Sucursal_Pereira,Sucursal_Santa Marta,Sucursal_Cartagena,Sucursal_Bucaramanga,Sucursal_Cucuta,Sucursal_Of. Principal,Sucursal_Duitama,Sucursal_Neiva,Sucursal_San Gil,Sucursal_Armenia,Sucursal_Buenaventura
-0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
