+++ pages/Inicio1.py
@@ -2,7 +2,7 @@ import dash_bootstrap_components as dbc
 from dash_labs.plugins.pages import register_page
 from dash import  dcc, html, Input, Output, callback
 
-import jose.graficas_pred_dane_ascension as josepred
+import components.graphs.graficas_pred_dane_ascension as josepred
